+++ logprep/connector/http/output.py
@@ -13,7 +13,7 @@ An example config file would look like:
     :linenos:
 
     output:
-      myhttpoutput:
+      my_http_output:
         type: http_output
         target_url: http://the.target.url:8080
         username: user
@@ -161,13 +161,9 @@ class HttpOutput(Output):
                 stats[key] = int(sample.value)
         return json.dumps(stats, sort_keys=True, indent=4, separators=(",", ": "))
 
-    def store(self, document: tuple[str, dict | list[dict]] | dict) -> None:
-        if isinstance(document, tuple):
-            target, document = document
-            target = f"{self._config.target_url}{target}"
-        else:
-            target = self._config.target_url
-        self.store_custom(document, target)
+    def store(self, document: str) -> None:
+        target, _, payload = document.partition(",")
+        self.store_custom(payload, target)
 
     def store_custom(self, document: dict | tuple | list, target: str) -> None:
         """Send a post request with given data to the specified endpoint"""
@@ -177,6 +173,9 @@ class HttpOutput(Output):
         elif isinstance(document, dict):
             request_data = self._encoder.encode(document)
             document_count = 1
+        elif isinstance(document, str):
+            document_count = document.count(";") + 1
+            request_data = document.replace(";", "\n")
         else:
             error = TypeError(f"Document type {type(document)} is not supported")
             self.metrics.number_of_failed_events += 1