+++ logprep/generator/http/input.py
@@ -17,9 +17,12 @@ import msgspec
 from attrs import define, field, validators
 from ruamel.yaml import YAML
 
+from logprep.generator.http.batcher import Batcher
 from logprep.generator.http.manipulator import Manipulator
 
 yaml = YAML(typ="safe")
+
+logger = logging.getLogger("Input")
 
 
 @define(kw_only=True)
@@ -81,7 +84,7 @@ class Input:
     MAX_EVENTS_PER_FILE = 100_000
 
     @cached_property
-    def _temp_dir(self):
+    def temp_dir(self):
         return Path(tempfile.mkdtemp(prefix="logprep_"))
 
     @cached_property
@@ -102,20 +105,19 @@ class Input:
         self.number_of_events = config.get("events")
         self.events_sent = 0
         self.batch_size = config.get("batch_size")
-        self.log = logging.getLogger("Input")
         self.log_class_manipulator_mapping: Dict = {}
         self.number_events_of_dataset = 0
         self.event_file_counter = 0
 
-    def reformat_dataset(self):
+    def reformat_dataset(self) -> None:
         """
         Collect all jsonl files of each event class and their corresponding manipulators
         and targets. The collected events will be written to one or multiple files containing
         the events and the target for the events.
         """
-        self.log.info(
+        logger.info(
             "Reading input dataset and creating temporary event collections in: '%s'",
-            self._temp_dir,
+            self.temp_dir,
         )
         start_time = time.perf_counter()
         events = []
@@ -126,7 +128,7 @@ class Input:
             self._populate_events_list(events, file_paths, log_class_config)
         if events:
             self._write_events_file(events)
-        self.log.info(f"Preparing data took: {time.perf_counter() - start_time:0.4f} seconds")
+        logger.info(f"Preparing data took: {time.perf_counter() - start_time:0.4f} seconds")
 
     def _retrieve_log_files(self, event_class_dir):
         """
@@ -151,7 +153,7 @@ class Input:
         config_path = os.path.join(event_class_dir_path, "config.yaml")
         with open(config_path, "r", encoding="utf8") as file:
             event_class_config = yaml.load(file)
-        self.log.debug("Following class config was loaded: %s", event_class_config)
+        logger.debug("Following class config was loaded: %s", event_class_config)
         event_class_config = EventClassConfig(**event_class_config)
         if "," in event_class_config.target_path:
             raise ValueError(
@@ -168,30 +170,33 @@ class Input:
             with open(file, "r", encoding="utf8") as event_file:
                 for event in event_file.readlines():
                     self.number_events_of_dataset += 1
-                    events.append(f"{log_class_config.target_path},{event.strip()}\n")
+                    events.append(f"{log_class_config.target_path},{event.strip()}")
                     if len(events) == self.MAX_EVENTS_PER_FILE:
                         self._write_events_file(events)
 
-    def _write_events_file(self, events):
+    def _write_events_file(self, documents) -> None:
         """
         Take a list of target and event strings and write them to a file. If configured the events
         will be shuffled first.
         """
         if self.config.get("shuffle"):
-            random.shuffle(events)
+            random.shuffle(documents)
         file_name = f"{self._temp_filename_prefix}_{self.event_file_counter:0>4}.txt"
-        temp_file_path = self._temp_dir / file_name
-        with open(temp_file_path, "w", encoding="utf8") as event_file:
-            event_file.writelines(events)
+        temp_file_path = self.temp_dir / file_name
+        batcher_config = {"batch_size": self.batch_size, "events": self.number_of_events}
+        logger.debug("Batcher config: %s", batcher_config)
+        batcher = Batcher(documents, **batcher_config)
+        with open(temp_file_path, "a", encoding="utf8") as event_file:
+            event_file.writelines(batcher)
         self.event_file_counter += 1
-        events.clear()
+        documents.clear()
 
     def load(self) -> Generator[List, None, None]:
         """
         Generator that parses the next batch of events, manipulates them according to their
         respective configuration and returns them with their target.
         """
-        input_files = [self._temp_dir / file for file in os.listdir(self._temp_dir)]
+        input_files = [self.temp_dir / file for file in os.listdir(self.temp_dir)]
         if self.config.get("shuffle"):
             random.shuffle(input_files)
         if self.number_of_events is None:
@@ -249,6 +254,6 @@ class Input:
 
     def clean_up_tempdir(self):
         """Delete temporary directory which contains the reformatted dataset"""
-        if os.path.exists(self._temp_dir) and os.path.isdir(self._temp_dir):
-            shutil.rmtree(self._temp_dir)
-        self.log.info("Cleaned up temp dir: '%s'", self._temp_dir)
+        if os.path.exists(self.temp_dir) and os.path.isdir(self.temp_dir):
+            shutil.rmtree(self.temp_dir)
+        logger.info("Cleaned up temp dir: '%s'", self.temp_dir)